🗑️ Remove Generated Proto Files
This commit deletes the generated protobuf files for the auth, major, school, and user modules, as well as the shared and validation proto files. This cleanup is part of a larger effort to streamline the codebase and ensure that only necessary files are maintained.

- Removed `authPb_grpc.pb.go`, `authPb.pb.go`
- Removed `majorPb_grpc.pb.go`, `majorPb.pb.go`
- Removed `schoolPb_grpc.pb.go`, `schoolPb.pb.go`
- Removed `userPb_grpc.pb.go`, `userPb.pb.go`
- Removed `database.pb.go`, `shared.pb.go`, `validation.pb.go`

This change helps in reducing clutter and improving maintainability of the codebase.